Working on commenting the codebase. About 40% done. Also, moved isTargetedByColor from PositionValidator to Positions
