+++ src/main/java/gg/w6/chesslib/model/Coordinate.java
@@ -106,19 +106,39 @@ public class Coordinate {
         return rank;
     }
 
+    /**
+     * Returns the index of the file of the coordinate
+     * <p>Semantically equivalent to <code>getFile().ordinal()</code></p>
+     * @return the index of the file of the coordinate
+     */
     public int getFileIndex() {
         return file.ordinal();
     }
 
+    /**
+     * Returns the index of the rank of the coordinate
+     * <p>Semantically equivalent to <code>getRank().ordinal()</code></p>
+     * @return the index of the rank of the coordinate
+     */
     public int getRankIndex() {
         return rank.ordinal();
     }
 
+    /**
+     * Returns the traditional string representation of a coordinate
+     * ex. "c4"
+     * @return the traditional string representation of a coordinate
+     */
     @Override
     public String toString() {
         return file.toString() + rank.toString();
     }
 
+    /**
+     * Returns <code>true</code> if <code>obj</code>, is a <code>Coordinate</code>, the files are equivalent, and the ranks are equivalent.
+     * @param obj the object with which to compare against
+     * @return <code>true</code> if equal
+     */
     @Override
     public boolean equals(Object obj) {
         return obj instanceof final Coordinate otherCoordinate
